+++ src/components/Header.tsx
@@ -4,7 +4,6 @@ import * as Icons from "./iconsComponents";
 
 const Header = () => {
   const { profile } = useAppSelector(state => state.user);
-  // const bellStyle = reminder ? "fill-red-500 stroke-red-500 animate-[bell_1.5s_ease-in-out_infinite]" : "";
 
   return (
     <div className="flex sticky top-0 items-center px-4 py-2 z-10 bg-bg-light border-b border-txt-grey">
@@ -21,9 +20,7 @@ const Header = () => {
         <span className="font-normal text-txt-main">Have a nice day</span>
       </p>
       <div className="ml-auto w-[34px] h-[34px]">
-        {/* <Link to="/reminder"> */}
         <Icons.Bell size={32} className="text-txt-main" />
-        {/* </Link> */}
       </div>
     </div>
   );
