+++ src/components/main.tsx
@@ -1,22 +1,44 @@
 /* eslint-disable prettier/prettier */
 /* eslint-disable no-nested-ternary */
+import { Form, Formik } from "formik";
+import { useState } from "react";
+
 import { useAppSelector } from "hooks/reduxHooks";
+import { setAuthHeader } from "redux/userOperations";
+import { IVacancy, useGetVacanciesQuery } from "redux/VacancyQueries";
 import * as Icons from "components/iconsComponents";
 import AddBtn from "components/addBtn";
-import { setAuthHeader } from "redux/userOperations";
-import { useGetVacanciesQuery } from "redux/VacancyQueries";
-
 import Loader from "components/ui/loader";
 import ShortNote from "components/notices/ShortNotice";
 import Header from "./Header";
 
+export interface Values {
+  text: string;
+}
+
 export default function Main() {
   const { token } = useAppSelector(state => state.user);
   setAuthHeader(token);
+  const [text, setText] = useState("");
+
+  const initialValues = {
+    text: "",
+  } as Values;
 
   const { data: response, isLoading, isError } = useGetVacanciesQuery();
 
-  const vacancies = response?.data?.filter(vacancy => vacancy.archived === false);
+  const vacancies = response?.data
+    ?.filter(vacancy => vacancy.archived === false && vacancy.companyName.toLowerCase().includes(text.toLowerCase()))
+    .sort((firstVacancy, secondVacancy) => secondVacancy.actions[0].date - firstVacancy.actions[0].date) as IVacancy[];
+
+  const handleChange = (event: React.FormEvent<HTMLFormElement> | React.ChangeEvent<HTMLInputElement>) => {
+    setText(event.currentTarget.value.toLowerCase());
+  };
+
+  const handelFormSubmit = (currentCurrency: Values): void => {
+    console.log(currentCurrency);
+    setText("");
+  };
 
   return (
     <div className="container mx-auto bg-bg-light">
@@ -26,28 +48,58 @@ export default function Main() {
         </div>
       ) : isError ? (
         <h2>ERROR</h2>
-      ) : !vacancies || !vacancies.length ? (
-        <>
-          <Header />
-          <div className="flex justify-center mt-24 px-4">
-            <Icons.Todos />
-          </div>
-          <p className="text-center text-txt-main text-xl px-10 pt-8">You do not have any job vacancies created yet</p>
-          <div className="flex justify-end mx-2 fixed bottom-32 right-8">
-            <AddBtn />
-          </div>
-        </>
       ) : (
-        <>
+        <div>
           <Header />
-          <div className="grid grid-cols-1 md:grid-cols-2 lg:grid-cols-3 mb-28 mt-5 items-center gap-4 px-4">
-            {vacancies && vacancies.map(vacancy => <ShortNote key={vacancy._id} shortVacancy={vacancy} />)}
-          </div>
+          <Formik initialValues={initialValues} onSubmit={handelFormSubmit} className="h-20 bg-bg-light">
+            <Form autoComplete="off" className="flex sticky top-0 left-0 right-0 mt-2">
+              <label htmlFor="company" className="w-full">
+                <input
+                  id="company"
+                  type="text"
+                  name="text"
+                  value={text}
+                  autoComplete="off"
+                  onChange={handleChange}
+                  placeholder="Enter company name"
+                  className="w-full h-8 px-2 py-1 outline-none rounded-md border "
+                />
 
-          <div className="flex justify-end mx-2 fixed bottom-32 right-8">
-            <AddBtn />
-          </div>
-        </>
+                {text === "" ? (
+                  <button type="button" className="absolute top-1 right-2 items-center">
+                    <Icons.Search />
+                  </button>
+                ) : (
+                  <button type="button" className="absolute top-1 right-2 items-center" onClick={() => setText("")}>
+                    <Icons.Close />
+                  </button>
+                )}
+              </label>
+            </Form>
+          </Formik>
+          {!vacancies || !vacancies.length ? (
+            <>
+              <div className="flex justify-center mt-24 px-4">
+                <Icons.Todos />
+              </div>
+              <p className="text-center text-txt-main text-xl px-10 pt-8">
+                You do not have any job vacancies created yet
+              </p>
+              <div className="flex justify-end mx-2 fixed bottom-32 right-8">
+                <AddBtn />
+              </div>
+            </>
+          ) : (
+            <>
+              <div className="grid grid-cols-1 md:grid-cols-2 lg:grid-cols-3 mb-28 mt-5 items-center gap-4 px-4">
+                {vacancies && vacancies.map(vacancy => <ShortNote key={vacancy._id} shortVacancy={vacancy} />)}
+              </div>
+              <div className="flex justify-end mx-2 fixed bottom-32 right-8">
+                <AddBtn />
+              </div>
+            </>
+          )}
+        </div>
       )}
     </div>
   );
